Struct symtern::adaptors::Inline

pub struct Inline<W> { /* fields omitted */ }

Interner adaptor optimized for short strings.

Inline will encode any string shorter than the symbol-ID type directly inside the symbol; strings of the same or greater size will be passed to the wrapped interner.

Simple benchmarks included with the crate indicate that this gives an approximately 6x (82%) speedup over the basic Pool for strings small enough to be inlined.

The downside to using this adaptor is its capacity, which is also "short": the current implementation uses one of the bits of a symbol's ID to mark it as containing an inlined string, which halves the number of addressable slots in the look-up table. But symbols with inlined values don't occupy any space in the pool, so this may be a net gain if you expect your input to be dominated by short strings.

use symtern::prelude::*;
use symtern::Pool;
use symtern::adaptors::Inline;

let mut pool = Inline::<Pool<str,u64>>::new();
let hello = pool.intern("Hello").expect("failed to intern a value");
let world = pool.intern("World").expect("failed to intern a value");

assert!(hello != world);

assert_eq!((Ok("Hello"), Ok("World")),
           (pool.resolve(&hello),
            pool.resolve(&world)));

// Since both "Hello" and "World" are short enough to be inlined, they
// don't take up any space in the pool.
assert_eq!(0, pool.len());

The internal Pack trait, which provides the inlining functionality, is implemented for u16, u32, and u64; it will be implemented for u128 as well when support for 128-bit integers lands.

Methods

impl<W> Inline<W>

pub fn new() -> Self where
W: Default,

Create a new, empty symbol pool