Struct symtern::adaptors::Luma

pub struct Luma<W> { /* fields omitted */ }

"Lifetime-safe" interner adaptor.

This adaptor's symbols are treated as references to their source pool — which, thanks to Rust's borrow-checker, prevents us from dropping the pool before any of its symbols.

To achieve this, we utilize interior mutability via RefCell (which allows us to intern more than one value at a time). Note that this is not a zero-cost abstraction: the price we pay for interior mutability is run-time borrow-checking! More specifically, certain borrow errors that would normally be caught at compile-time...

use symtern::prelude::*;
use symtern::Pool;

let mut pool = Pool::<str, u32>::new();
let x = pool.intern("foo").expect("failed to intern a value");
let foo = pool.resolve(x).expect("failed to resolve the value we just interned");
assert_eq!("foo", foo);

let _ = pool.intern("bar").expect("failed to intern a value"); //~ ERROR cannot borrow `pool` as mutable because it is also borrowed as immutable

...become run-time errors when using the Luma adaptor:

use symtern::prelude::*;
use symtern::Pool;
use symtern::adaptors::Luma;

let pool = Luma::from(Pool::<str, u32>::new());
let x = pool.intern("foo").expect("failed to intern a value");
let foo = pool.resolve(x).expect("failed to resolve the value we just interned");
assert_eq!("foo", &*foo);

let _ = pool.intern("bar").expect("failed to intern a value"); //~ PANIC already borrowed: BorrowMutError

Methods

impl<W> Luma<W>

pub fn new() -> Self where
W: Default,

Create a new, empty Luma instance.