Module symtern::prelude

Items intended for glob-import.

Symtern takes a conservative stance on what names should will be brought into scope when you glob-import this module's contents: only traits necessary for method resolution are included, and the names of all re-exported traits are prefixed with Symtern to reduce the likelihood of name collisions.

This stance may change in future versions, but for now must write your own use statements for any other Symtern types you wish to use.

Reexports

pub use traits::Len as SymternLen;

pub use traits::Intern as SymternIntern;

pub use traits::Resolve as SymternResolve;

pub use traits::ResolveUnchecked as SymternResolveUnchecked;