Struct symtern::Pool

pub struct Pool<T: ?Sized, I = usize> where
    T: ToOwned + Eq + Hash,
    T::Owned: Eq + Hash,
    I: SymbolId,  { /* fields omitted */ }

Simple hash-based interner generic over both the type of interned values and the type used to represent symbol IDs.

Pool can intern any type that implements ToOwned, Eq, and Hash, where its owned type (ToOwned::Owned) also implements Eq and Hash.

use symtern::prelude::*;
use symtern::Pool;

#[derive(Clone, Eq, PartialEq, Hash)]
struct WibbleWobble {
    whee: Vec<u32>
}

let mut pool = Pool::<_,u8>::new();
assert!(pool.intern(&WibbleWobble{whee: vec![1, 2, 3, 4, 5]}).is_ok());

Methods

impl<T: ?Sized, I> Pool<T, I> where
T: ToOwned + Eq + Hash,
T::Owned: Eq + Hash,
I: SymbolId,

pub fn new() -> Self

Create a new, empty Pool instance.