Trait symtern::traits::Intern

pub trait Intern {
    type Input: ?Sized;
    type Symbol: Symbol + Symbol;
    fn intern(self, value: &Self::Input) -> Result<Self::Symbol>;
}

Primary interface for interner implementations.

In order to allow for implementations that require a lifetime parameter, and to abstract over mutability requirements, this trait's methods take self by value; for a given type T, the trait should implemented for either &'a T or &'a mut T.

Associated Types

type Input: ?Sized

Type of value accepted by intern.

type Symbol: Symbol + Symbol

Type used to represent interned values.

Required Methods

fn intern(self, value: &Self::Input) -> Result<Self::Symbol>

Fetch the symbol that corresponds to the given value. If the value does not map to any existing symbol, create and return a new one. This method may return an error if the interner encounters any error while storing the value.

let symbol = match some_interner.intern("Rosebud") {
    Ok(sym) => sym,
    Err(err) => return Err(MyErrorType::from(err)),
};