Trait symtern::traits::Len

pub trait Len {
    fn len(&self) -> usize;

    fn is_full(&self) -> bool;

    fn is_empty(&self) -> bool;
}

Trait for use with interners that can report the number of values they contain.

Required Methods

fn len(&self) -> usize

Fetch the number of values contained in the interner.

fn is_full(&self) -> bool

Check if the number of interned symbols has reached the maximum allowed.

fn is_empty(&self) -> bool

Check if the interner is "empty", i.e. has zero stored values.